practical 11 handled exceptions

+++ Practical11/Test2/Models/Services/StudentService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using Test2.Models.Entities;
 using Test2.Models.Repositories;
 
@@ -20,22 +21,58 @@ namespace Test2.Models.Services
 
         public Student GetById(int id)
         {
-            return _studentRepo.GetById(id);
+            if (id <= 0)
+            {
+                throw new ArgumentException("Id must be greater than zero.");
+            }
+
+            var student = _studentRepo.GetById(id);
+
+            if (student == null)
+            {
+                throw new KeyNotFoundException($"Student with id {id} not found.");
+            }
+
+            return student;
         }
 
         public void Create(Student student)
         {
+            if (student == null)
+            {
+                throw new ArgumentNullException(nameof(student), "Student cannot be null.");
+            }
+
+            if (string.IsNullOrWhiteSpace(student.FullName))
+            {
+                throw new ArgumentException("Student name is required.");
+            }
 
             _studentRepo.Add(student);
         }
 
         public void Update(Student student)
         {
+            if (student == null)
+            {
+                throw new ArgumentNullException(nameof(student), "Student cannot be null.");
+            }
+
+            if (string.IsNullOrWhiteSpace(student.FullName))
+            {
+                throw new ArgumentException("Student name is required.");
+            }
+
             _studentRepo.Update(student);
         }
 
         public void Delete(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("Id must be greater than zero.");
+            }
+
             _studentRepo.Delete(id);
         }
     }
